Consolidate API functions
