+++ web/api/recipients.js
@@ -1,5 +1,5 @@
 import { kv } from '@vercel/kv';
-import { requireAdmin } from './auth.js';
+import { requireAdmin } from '../utils/auth.js';
 
 // KV Helper functions for Recipients
 async function getRecipientsFromKV() {
