+++ web/components/config/config.js
@@ -1,7 +1,7 @@
-export const API_STATUS = '/api/status';
-export const API_RUNS = '/api/runs';
-export const API_RUN = '/api/run';
-export const API_LOGS = '/api/logs';
-export const API_ARTIFACT = '/api/artifact';
+export const API_STATUS = '/api/github?action=status';
+export const API_RUNS = '/api/github?action=runs';
+export const API_RUN = '/api/github?action=run';
+export const API_LOGS = '/api/github?action=logs';
+export const API_ARTIFACT = '/api/github?action=artifact';
 export const API_LOGIN = '/api/login';
-export const API_USER_LOGIN = '/api/user-login';
+export const API_USER_LOGIN = '/api/login';
