+++ web/test/api-artifact.test.js
@@ -14,14 +14,14 @@ function makeRes() {
 
 
 async function load(adminReturn) {
-  const mod = await import('../api/artifact.js?' + Date.now());
+  const mod = await import('../api/github.js?' + Date.now());
   mod.__setRequireAdmin(() => adminReturn);
   return mod;
 }
 
 test('rejects non-GET', async () => {
   const res = makeRes();
-  const { default: handler } = await import('../api/artifact.js?' + Date.now());
+  const { default: handler } = await import('../api/github.js?' + Date.now());
   await handler({ method: 'POST' }, res);
   assert.equal(res.code, 405);
   assert.equal(res.sent, 'Method Not Allowed');
@@ -30,7 +30,7 @@ test('rejects non-GET', async () => {
 test('requires admin', async () => {
   const { default: handler } = await load(null);
   const res = makeRes();
-  await handler({ method: 'GET', query:{} }, res);
+  await handler({ method: 'GET', query:{ action:"artifact" } }, res);
   assert.equal(res.code, null);
   assert.equal(res.sent, null);
 });
@@ -38,7 +38,7 @@ test('requires admin', async () => {
 test('missing id', async () => {
   const { default: handler } = await load(true);
   const res = makeRes();
-  await handler({ method: 'GET', query:{} }, res);
+  await handler({ method: 'GET', query:{ action:"artifact" } }, res);
   assert.equal(res.code, 400);
   assert.equal(res.sent, 'Missing id');
 });
@@ -49,7 +49,7 @@ test('fetch error propagates', async () => {
   process.env.GH_REPO = 'r';
   process.env.GH_TOKEN = 't';
   const res = makeRes();
-  await handler({ method: 'GET', query:{ id: '1' } }, res);
+  await handler({ method: 'GET', query:{ action:"artifact", id: '1' } }, res);
   assert.equal(res.code, 404);
   assert.equal(res.sent, 'oops');
 });
@@ -61,7 +61,7 @@ test('successful fetch sends zip', async () => {
   process.env.GH_REPO = 'r';
   process.env.GH_TOKEN = 't';
   const res = makeRes();
-  await handler({ method: 'GET', query:{ id:'1' } }, res);
+  await handler({ method: 'GET', query:{ action:"artifact", id:'1' } }, res);
   assert.equal(res.headers['Content-Type'], 'application/zip');
   assert.deepStrictEqual(res.sent, buf);
 });
