+++ web/test/api-auth.test.js
@@ -14,7 +14,7 @@ test('missing secret returns 500', async () => {
   const req = { headers: {} };
   const res = makeRes();
   process.env.JWT_SECRET = '';
-  const { requireAdmin } = await import('../api/auth.js?' + Date.now());
+  const { requireAdmin } = await import('../utils/auth.js?' + Date.now());
   assert.equal(requireAdmin(req, res), null);
   assert.equal(res.code, 500);
   assert.deepEqual(res.data, { message: 'Server configuration missing' });
@@ -24,7 +24,7 @@ test('missing auth header returns 401', async () => {
   const req = { headers: {} };
   const res = makeRes();
   process.env.JWT_SECRET = 's';
-  const { requireAdmin } = await import('../api/auth.js?' + Date.now());
+  const { requireAdmin } = await import('../utils/auth.js?' + Date.now());
   assert.equal(requireAdmin(req, res), null);
   assert.equal(res.code, 401);
 });
@@ -35,7 +35,7 @@ test('invalid token returns 401', async () => {
   const req = { headers: { Authorization: 'Bearer xx' } };
   const res = makeRes();
   process.env.JWT_SECRET = 's';
-  const { requireAdmin } = await import('../api/auth.js?' + Date.now());
+  const { requireAdmin } = await import('../utils/auth.js?' + Date.now());
   assert.equal(requireAdmin(req, res), null);
   assert.equal(res.code, 401);
 });
@@ -46,7 +46,7 @@ test('non admin role returns 403', async () => {
   const req = { headers: { Authorization: 'Bearer xx' } };
   const res = makeRes();
   process.env.JWT_SECRET = 's';
-  const { requireAdmin } = await import('../api/auth.js?' + Date.now());
+  const { requireAdmin } = await import('../utils/auth.js?' + Date.now());
   assert.equal(requireAdmin(req, res), null);
   assert.equal(res.code, 403);
 });
@@ -58,7 +58,7 @@ test('valid admin returns decoded object', async () => {
   const req = { headers: { Authorization: 'Bearer good' } };
   const res = makeRes();
   process.env.JWT_SECRET = 's';
-  const { requireAdmin } = await import('../api/auth.js?' + Date.now());
+  const { requireAdmin } = await import('../utils/auth.js?' + Date.now());
   const result = requireAdmin(req, res);
   assert.deepEqual(result, decoded);
   assert.equal(res.code, null);
